A bit of an offline-hacking mingled commit.
Add an option (--html-shortenlists) for the JavaScript list-shortening
stuff.

Add help text for all options.

Add a parentMod attribute to Documentable, and maintain it in
astbuilder.

Do subclass updating and borrrowing docstrings from superclasses or
interfaces much earlier, meaning that the default
finalStateComputation is much simpler (and will probably be trivial
pretty soon).

Make sourceHref computation a ASTBuilder method.

Add a long-ish comment to model.py about how Documentable is overly
generic.

Allow deletion of a docstring in the server mode, and go looking for
docstrings in superclasses and interfaces when this is done.