Construct plaintext directly in consolidated field conversion fallback
If epydoc finds a field like "parameters" that is actually several
fields combined into one, it will try to split up that consolidated
field. If parsing of the field's body fails, as a fallback it will
register the field as a new field type and preserve the body as
plain text.

In that scenario, the plain text version of the body was produced
by calling the parse() dispatcher function with 'plaintext' as its
markup argument. This commit replaces that call by a direct call
to the ParsedPlaintextDocstring constructor.

I added two doctest cases since we had zero coverage of field parsing.