Reshuffle implementation of visit_Assign()
This method is implemented to handle assignments of all results.
One part of the method handled such assignments in a module scope,
the other part in a class scope. Both parts were given dedicated
methods.

After that split, no early exits are needed anymore in visit_Assign().
As a result, we can just call the superclass implementation directly
instead of wrapping it in a lambda to avoid duplicate code.

Two redundant None checks were removed: isinstance() accepts None
values and will return False for any type checked except NoneType,
so a True/not-None check prior to isinstance() is not needed.