Move function to add attribute into ASTBuilder
This was an internal function of ZopeInterfaceModuleVisitor's
_handleAssignmentInClass() method. Moving it to ASTBuilder makes
it availabe for other code as well.

I also renamed the function from pushAttribute() to addAttribute()
because it only pushed the Attribute to trigger the side effects
of the push: it popped it back off the stack before returning.

The relevant side effects were adding the new attribute to the
system and setting parentMod, so those two things are now done
inside addAttribute() and the push;pop sequence is avoided.