Create Attribute for class variables in astbuilder's visitor
Instead of creating the Attribute when we have a lot of info to store,
we now create it as soon as we can store something. Later visitor
methods can then add information to this Attribute object.

This makes it possible to combine information from different sources,
such as type annotations and inline docstrings.