Allow documentation of instance variables using inline docstrings
The docstrings are placed in a method, below a assignment that assigns
to the instance variable. This was not part of PEP 224, but it is
a feature that the other documentation extraction tools have.

Typically these docstrings are placed in the __init__() method, but
I figured it wasn't necessary to restrict it on method name, since
there is a low chance of false positives and in some cases it might
be useful to support initialization from other methods. If this turns
out to be a problem in practice, we might have to make it stricter.