Make node2dottedname() far more selective
The previous implementation leads to a lot of assignments being
treated as aliases when they shouldn't be. That was mostly harmless
up to now, but with the upcoming changes it will break things.

This commit more or less restores the implementation from before
the 'compiler' to 'ast' migration, but with the code updated to
work with 'ast'.