Fix class variable overriding instance variable
A class variable should override a variable with the same name from
a higher scope. However, if a variable seems to be both an instance
and class variable, it is likely an instance variable with a default
set at the class level, so we should re-use the existing Attribute
instead of replacing it with a new one.