Modernize astbuilder.parseFile()
Read file as binary. This is accepted by ast.parse(). Previously the
mode "U" (universal newlines) was used, but this is deprecated since
Python 3.4 and scheduled for removal in 3.9. While universal newlines
is still present as a feature, the syntax to enable it is different
for Python 2 vs 3, so reading as a binary is easier.

Use open() in a 'with' statement to ensure that the file is properly
closed even when reading fails. It's also shorter.

The docstring referred to the 'compile' module and the 'MyTransformer'
class, both of which are no longer used since we switched to parsing
with the 'ast' module. This docstring and also the one of parse()
were updated.