Support combination of @type with variable docstring
While this is not a combination that is likely to be used much,
it is still good if it works as expected.

The main reason for this change is to allow for type information
to be extracted from other sources than @type tags, such as type
annotations and attr.ib declarations.