Fix and expand keyword list for colorizing Python code
The original list was broken because a keyword at the end of the line
would be merged with the keyword at the start of the next line, since
there was no whitespace between them inside the string, only outside.

I also expanded the list to include keywords added to Python after
this code was originally written.