Decided to keep the HTML to Stan parsing for ReStructuredText
While it would be more efficient to have a dedicated translator from
Docutils nodes to Stan, there are a lot of different node types, so
writing such a translator is not a small task.

For documentation formats that might be added in the future, it is
also useful to have the HTML to Stan parsing as an option, even if
it is only to reduce the time needed to create a proof of concept.