Only return Tag objects from to_stan() methods
Previously, we returned various types that are allowed as Tag
children. Having a single return type simplifies introspection
and will make things easier when we start adding type annotations
to pydoctor's code in the future.

The stripping of trailing newlines in the HTML parsing default
implementation of to_stan() was dropped because the unit tests
did not require it and web browsers don't care much about
whitespace.

The epytext formatter handled 'html' nodes, but its parser never
produces those, so I removed the handling. It also handled 'tag'
and 'arg' nodes, which are produced by the parser, but only inside
'fieldlist' nodes. Since those should be removed from the tree
before formatting, 'tag' and 'arg' are now also treated as errors.