Improve the look of the plaintext formatter output
The <pre> tag is styled by bootstrap in a way that looks good for
inline code fragments, not for entire docstrings and especially
not in the summary tables.

Instead of using the tag, the plaintext formatter now outputs
a paragraph with a 'pre' CSS class attached. This makes the full
version of the docstring preformatted but without it being visually
framed, while the summary version will discard the paragraph and
just be an actual plain string.

I also dropped the newlines inserted at the start and end of the
content, since I don't see any reason for those to be there.