Fall back to plaintext formatter if fancier formatter fails
We had several error handlers calling boringDocstring() to fall back
on a plain presentation. This created a lot of different flows through
the code, which makes it hard to reason about it.

Instead of using boringDocstring(), we now fall back on the plaintext
formatter. This produces very similar output, but we can resume on
the happy path instead of bailing out early, so there are fewer flows
to deal with.

An additional advantage is that we no longer skip field presentation
when the main docstring has errors. This is useful now that we have
more sources of field documentation than just the main docstring.