Resolve to-do comments about System construction
Passing the 'options' object is a simple and safe change.

I decided not to implement the other comments:

Putting the call to fetchIntersphinxInventories() in System.__init__()
goes against the practice of dependency injection. In this particular
case it would complicate unit testing, since we'd have to provide
a dummy cache.

Passing the project name to System.__init__() is not possible, since
the project name is autodetected when the user doesn't provide it,
but that autodetection requires System.rootobjects to be populated.