Split SphinxInventory into a reader and writer class
The only overlap between the two was in the logging.

This solves the problem of not being able to pass the project name
to the inventory when a System is first created: since the reader
doesn't need the project name, we don't have to pass it and by the
time we create a writer we have autodetected the project name.