Convert some of the epytext tests from doctest to functions
Plain Python code is easier to edit and is understood by tools like
pyflakes. Since pytest shows detailed information for failed asserts
we don't need doctest to provide that.

I did leave some tests with multi-line outputs as doctests for now,
as those are slightly more readable when there is an output mismatch.