Set line number for attributes on assignment
If an attribute is first mentioned in the docstring of its parent,
an object is created for it without line number information. If later
an assignment is done to the attribute, the line number wasn't set
because the object for the attribute already existed. To fix this,
the setting of the line number was taken out of the object creation
and moved to the assignment handlers.

An alternative approach would be to remember the line number from
the docstring tag. That approach does not conflict with this change,
since the first encountered line number is the one remembered.