Use OrderedDict to keep track of objects in a System
This was done with two variables until now: 'allobjects' as
an unordered dictionary and 'orderedallobjects' as a list of
the contents in order of adding.

I removed some redundant calls to fullName(): computing this
multiple times in a single function is just wasteful. I also
removed the unused return value from handleDuplicate().