Apply docstring inheritance for interface attributes
The inheritance for interface members is implemented separately from
class members, so it had to be updated to make attributes inherit too.

There was a bug in ASTBuilder.addAttribute() which always created
a model.Attribute instead of using the system.Attribute factory.
Until now this was harmless, since that factory was never overridden,
but now we do override it in ZopeInterfaceSystem.