Improve reporting of epydoc errors
Avoid reporting the same errors twice by filtering out duplicates.
It would be more efficient to simply not attempt to parse a docstring
again if it failed the first time.

Report the file name if it's known (in practice, always except for
unit tests) instead of the dotted object name. This makes it easier
for the user to find the problem in the source code and correct it.

All objects should now have a line number and even if it's not filled
in for some reason, it will be line 0 rather than undefined, so there
is no need for getattr().

Note that the actual line numbers are inaccurate: it uses the
documented object's line number as the base rather than the docstring's
line number. But that's something to solve in another commit.

Improve type checks: 'err' could be a unicode string in Python 2.7,
while if 'err' is not a string, we now verify that it is a ParseError.

Include the context that epydoc provides (by calling descr()) rather
than dumping the entire docstring at high verbosity levels.