Improve error reporting if @implementer arg is not an interface
The current code will patch the base class info to make it into
an interface. That is the right solution if we're likely to generate
false positives, but if that is unlikely, it would be better to
report the issue and ignore the non-interface argument.

Lowering the message threshold to the default value of 0 means we'll
get reports now at the default verbosity level if patching happens.

I made the unit test verify the current behavior, not because I'm
certain that is the correct behavior, but because we want to be
alerted if the behavior changes unintentionally.