Remove ability to use "_()" on zope.interface/schema docstrings
This feature was originally added in commit 283fa471. However, it is
not used in Twisted and I cannot find any other package that is using
it either.

Localizing a docstring definition has some problems:
- it requires the i18n system to be up and running really early
- changing the locale requires reloading the modules
- in the case of webapps, a single server-selected locale will be
  used for all clients
- it doesn't work for regular docstrings, which must be string
  literals, so some __doc__ values will be localized and others not

An application that wants to localize field descriptions should
probably do so as part of the presentation instead of the definition.

All in all, I think this is not a pattern we must support.

Note that the implementation of extractStringLiteral() got broken in
the migration to the "ast" module, specifically in commit 5cf66a9b:
the handling of ast.Call does not check that the function name is "_"
and the handling of ast.Name does not make sense. However, it did not
break the unit test, so no-one noticed.

By the way, raising an error for non-strings in extractStringLiteral()
is also wrong, since this can be triggered by bad input rather than
bugs in pydoctor. But that is going to be addressed in a separate
commit in the near future.