Support inline docstrings for InterfaceClass instantiations
Handling of instantiations of zope.interface.interface.InterfaceClass
subclasses seems to be only partially implemented: the docstring was
hardcoded to "...", relevant arguments such as 'bases' and '__doc__'
are ignored and the message "new interface" was logged at the default
verbosity level.

This commit removes the hardcoded docstring and applies an inline
docstring following the assignment to the new interface.

Full support would mean also processing the 'bases' and '__doc__'
arguments, but that is a non-trivial amount of work that I'm not
motivated to do right now since I don't use it, Twisted doesn't
use it and I'm not aware of anyone else who would be using it.