Ignore documentables inside functions
These cannot be called directly, so they should not be considered to
be part of the API.

In theory they could be returned from the function and then exposed,
but I don't think our aliasing support is powerful enough to support
that anyway.

Note that documented local variables were never included in the
output, but I added them to the test case to make sure that won't
happen in the future either.

Closes #195