Use start of the decorator list as lineno for function/class
Before Python 3.8, the lineno of the function/class node itself was
set to the start of the decorator list, but in 3.8 that changed to
the line of the `def`/`class` statement.

While I think the new behavior is more accurate, for the purpose of
linking to a line in the source code we want the old behavior,
since the user will want to see the definition in its entirity.