Do not warn about relative imports
Python 2 supported implicit relative imports and pydoctor warned
when those were used. While I can imagine that a project wouldn't
want implicit relative imports, I'm not sure it's pydoctor's job
to warn about them. In any case, Python 3 only allows explicit
relative imports, which so there is nothing to warn about anymore.

In the migration to Python 3, the test data was updated to use
an explicit relative import, but the test still checked whether
a warning was printed and because the warning was incorrectly
also printed for explicit relative imports, the test passed.