Remove support for relative imports without 'from' keyword
This is no longer supported in Python 3 and was a bad idea in
Python 2, so I don't think there is any point in supporting it.

I'm not sure the code was correct to begin with. The fact that
no test cases failed when I removed the code only adds to those
doubts.