Use Twisted trunk both as input and for execution of twisted-apidoc
Previously, Twisted trunk was used as the input, but the execution was
done using the latest Twisted from PyPI.

Twisted trunk now contains Python 3 code that cannot be parsed by
Python 2.7, so we need to run pydoctor under Python 3. But the
twisted-apidoc tox environment makes use of a release module from
Twisted that isn't Python 3 compatible in the latest Twisted release.

The solution I picked is to put the Twisted trunk Git clone in
PYTHONPATH, such that the twisted-apidoc environment executes on
Twisted trunk.