Fix cross-references to builtin types in standard library
In Python 3, the exception types exist in the 'builtins' module instead
of the separate 'exceptions' module that existed in Python 2. The code
dealt with this by assigning the 'builtins' module to the variable
named 'exceptions'.

While this avoided crashes, it broke the check whether a name belongs
to an exception: builtin types such as 'str' would exist in the
module referenced by the 'exceptions' variable and would therefore
be linked to the documentation page for builtin exceptions.

The solution is to check instead whether the name represents
a subclass of BaseException.