Use caplog fixture to test IntersphinxCache logging
Since pytest knows how to capture logging from the standard library,
we don't have to write a custom logger for testing.

The check was also made more strict: instead of checking whether any
exception is logged, we now check for a specific custom exception.
This avoids the risk of an unrelated exception making the test pass.