Remove PersistentBytesIO test helper class
When BytesIO is closed, its buffer is discarded. PersistentBytesIO
overrides the close() method to avoid this, but the only reason
the close() method was called is that the test case wrapped it in
contextlib.closing().

By using @contextmanager instead, we no longer need the custom
BytesIO subclass. And the resulting code will be easier to
annotate.