Use fixtures to set up SUT in Sphinx inventory tests
This creates more consistency among the test cases. But the main
motivation is that it should be easier to add type annotations now.

We could check whether the recorded messages list is empty instead
of using separate '_nolog' fixtures. But if such a test would fail,
it would be more work to track down where the unexpected message
is logged from.