Do not set Module.ast attribute in astbuilder tests
This attribute was used to pass the AST object from the fromText()
test helper function  to specific test cases. Since this attribute
is not defined in the Module class, mypy will flag assignments to
it as an error.

I split off the code in fromText() that uses the AST into a new
fromAST() test helper function, so test cases that need access to
the AST can create it locally before calling the helper.