Do not wrap all field contents in <code>
The parameter descriptions should not be formatted as code, since they
are usually in natural language and if they do contain code, epytext
tags can be used to mark segments as code.

This reverts commit dba57244 ("adding <code> everywhere") and updates
the unit tests to make them pass again after the revert.

Looking at Twisted's docstrings, `@type` and `@rtype` use `C{}` or
`L{}` if they contain a type name. So I think it makes sense to not
expect the value of a type field to be auto-wrapped in <code>,
unlike an annotation, which originates from the actual code.