Add CSS for syntax highlighting and doctest blocks
The styling for this got lost in the big makeover of 2015, that is,
commit 620c5070 ("Add a new style based on Bootstrap and remove the
old JavaScript.")

I tried resurrecting the old CSS that originated from epydoc, but
it didn't match very well with the rest of our styling. So I made
new rules instead, using some existing colors from our CSS and
using GitHub and Kate as inspiration for new styles.