When '@param' and '@type' names include asterixes, warn and strip them
We can't match the annotations to docstring fields if the names don't
align. Since the asterixes are redundant and less frequently used in
Twisted's documentation, asterix-less is now the official syntax.

In theory we could strip the asterixes without warning, but for the
sake of consistency we'd rather not have them in the docstrings.