Report docstring field issues on source location
If a docstring is inherited, the object we're processing does not
match the position in the source code that we should be reporting
the issue on.

We still have the problem that issues in inherited docstrings are
reported multiple times, but at least the same correct location is
repeated now, which makes it easier to spot duplicates.