Do not print 'XXX' on duplicate `@return` or `@rtype`
This message is probably a leftover debug message. In any case, it is
not very informational to the user.

It might be useful to warn on duplicate fields for fields where there
is no need to include them multiple times, but that should be done in
a more consistent way. For example, documenting the same parameter more
than once could also be warned about. I created #229 as reminder.