Convert format_desc_list() and format_field_list() to generators
The early outs for empty lists were unnecessary and it is doubtful
they matter much in performance.

Instead of adding a list to the table contents, we now add individual
rows, but since Stan flattens everything in the end, the output will
be identical.

Overall, we create fewer temporary objects and use fewer lines of code.