Require that extract_fields() is only called for documented objects
This was already the case, but extract_fields() didn't rely on this,
which led to extra work. In particular, it would call get_docstring()
which in theory could return inherited docstrings and we wouldn't
want to inherit fields in general.

In practice, since the given object is documented, get_docstring()
returned that object's own docstring. So there was no bug here, but
there was needless complexity.