Make name of resolve_identifier_xref() parameter consistent
In practice we only pass positional arguments, so the fact that the
abstract base class and implementation class used different names
didn't break anything at runtime. But it's still incorrect and it
will cause problems once we start generating our own API docs.

In addition to that, 'fullID' not actually being the full ID and
therefore there being a 'fullerID' as well was just confusing.