Remove stdlib_doc_link_for_name()
This function attempts to construct a URL to the documentation for
a name in the Python standard library. However, there is no simple
algorithm to do that reliably, since there are too many things
documented in places that don't have a straightforward relation
to their full name. See #212 for some examples.