Fix file names for function pages
This got broken when introducing the 'url' property in commit 106621e8
("Deduplicate code that computes the relative URL of HTML output").
The link() function in pydoctor.templatewriter.util wasn't used to
produce links, but to produce an output path instead. So using the
relative URL there led to file names with hash marks in them.

In the fix, I opted for using Twisted's FilePath over the quote()
function from urllib.parse: local file paths are not URLs, so
escaping them as such might cause mismatches. At the same time,
it is a good idea to avoid accidentally writing outside our
output base, which FilePath protects against.