Do not fail twisted-apidoc tox env on non-empty log
In practice, when the log is non-empty it's usually due to a mistake
in one of Twisted's docstrings, so failing pydoctor's system test
for reporting errors is shooting the messenger.

But the bigger problem is that we cannot see the difference between
pydoctor crashing, which we always want to be alerted to, and it
logging warnings that might be an indication of a problem in
pydoctor.