Remove continue-on-error from system test workflow
When continue-on-error is enabled and a step fails, that failure does
not prevent the overall test status from turning green. So unless you
dig deeper into the reporting, you might think everything is well when
there is in fact a failing test.

Now that the twisted-apidoc tox env no longer considers a non-empty
log to be an error, all non-zero exit codes are real errors, so we
don't need to treat this step differently anymore.