Fix parsing of Sphinx inventory lines that contain spaces
The previous implementation assumed that all extra spaces belonged
to the 'display name' column at the end, but that is not correct
since the 'name' column (first column) can contain spaces as well.

The test case is based on fragments from the Python documentation,
so this is not just a theoretical issue.