Remove check for builtins from Module._localNameToFullName()
The builtins don't exist at the module level, as can be seen when you
attempt to import a builtin like `int` from an arbitrary module.

A module does have a __builtins__ dictionary, but those names form
a scope above the module level; they're not names in the module.

Since the following `else:` case returns the same value, this removal
will not affect the behavior in any way.