Refactor part() slot rendering
A new method namespace() does the actual rendering of the full name,
while part() composes the final output.

There is a subtle difference in behavior: part() only checked whether
the immediate parent is an __init__ module, while namespace() checks
at every level. In the case of nested class definitions, the new
behavior is correct while the old behavior is not.