Clean up code that handles re-exports of imported names
There were a lot of checks that were redundant or that were needlessly
done inside the loop.

The overwriting of the 'mod' variable was particularly nasty, since it
would put a package there if we're importing from an __init__ module,
and then following names would be looked up in the package instead.