Remove check that won't add object to parent if full names are equal
This check got added without explanation in c17fe008. That commit
was part of the Python 3 porting effort, but this particular change
doesn't seem to be directly related to that.

Maybe it helped unit tests pass at the time, but in that case it was
likely masking a bug elsewhere rather than fixing things, since the
parent's full name should never be equal to the child's full name.