When importing from a package, do name lookups from the package level
This fixes a regression introduced when getProcessedModule() started
returning the __init__ module.

A name imported from a package can be a module/subpackage, or a name
from the __init__ module. The former only works when we look up the
name in the package, while the latter will work whether we look up
in the package or the __init__ module, since the package will delegate
to the __init__ module when necessary.