Add type annotations to 'driver' module
The type check for classes looked up by name was moved into the
findClassFromDottedName() function, which means the type of classes
selected using the `--html-writer` option will now also be checked.

To preserve the current behavior, I implemented a custom subclass
check that will accept any class that implements the same public
methods. I also made some of the implementation methods private
to exclude them from the required interface, such that the
InMemoryWriter test class is considered to comply with the
TemplateWriter interface.

Perhaps we could just drop the `--html-writer` option: Twisted
doesn't use it and for people who want to customize report
generation in the future I'd prefer to support exporting the
docstrings and model info to JSON.