Improve annotation of SphinxInventory.update()
Originally this was annotated as a Mapping, but System passes an
IntersphinxCache instance, which doesn't implement the full mapping
interface. However, annotating as IntersphinxCache complicates
testing, where we pass dictionaries.

The solution is to define a Protocol which only includes the get()
method, such that both IntersphinxCache and 'dict' implement it.