Move anchors up to compensate for sticky navbar
When navigating to an anchor, browsers will typically put it at the
top of the page, but they don't take the sticky navbar into account,
so the thing the user was looking for ends up covered by the navbar.

Since anchors are invisible, we can move their location without
problems. However, they don't have an actual location by default.
But if we provide invisible content, they do, and the trick works.

Fixes #270.