Work around incorrect name lookups on cyclic imports
If a module hasn't been fully processed yet, looking up a name using
expandName() will return the wrong result, leading to worse output
than using the fallback. In particular, if expandName() cannot find
the name because it hasn't been processed yet, it will return the
unqualified name instead of the full name.

I'm not happy with this solution though, since the output becomes
unpredictable now: the order in which modules are processed is
random (probably depends on the hash seed). This can be demonstrated
by reverting the fix and running the test_cyclic_imports() test
several times: it will fail on the first assert half of the time
and on the second assert half of the time.