Remove tracing of imported names across module boundaries
This feature doesn't work reliably when the code we're analyzing
contains cyclic imports. And to make things worse, the way in which
it fails depends on the order in which modules are processed, which
isn't predictable and can vary between runs.

As projects adopt type annotations, it becomes increasingly common to
have cyclic imports of types, within `if TYPE_CHECKING:` blocks.
So I think this is something that pydoctor must be able to deal with
in a predictable manner.

Another motivation for removing the feature is that I have some doubt
about its usefulness: importing a name that is itself imported from
another module is bad practice, except when the name is explicitly
re-exported via `__all__`. But in the case of a re-export, the
documentation will be using the re-export location, so we don't need
to trace it back to the original implementation location.