Warn about non-existing parameter in 'type' field
We already had this warning for the 'param' field, but documenting
a type for a parameter that does not exist is equally worthy of
receiving a warning.

This also solves an issue where a 'type' field for a non-existing
parameter that precedes the 'param' field hides the warning for
that 'param' field, since having an entry in 'self.types' is how
we decide that a parameter exists.