Do not record calls to System._warning() as violations
This method is used to report situations in which pydoctor doesn't
understand the analyzed code, which can be because the code is
broken or because pydoctor isn't smart enough. In the former case,
code quality tools will report the issue as well, so we don't have
to. In the latter case we shouldn't fail the build, because it's not
a problem that the user can fix.

See #302 for details.