Use a property's `return` and `rtype` fields when available
When a property is documented like a function, use the `return` field
as the docstring body if the actual body is empty and use the `rtype`
field to document the type.

I think this approach is not as elegant as using the docstring body
plus type annotations, but in Twisted there are a few occurrences
of this pattern. And we like to have a way to override the type
documentation anyway.