Avoid empty summary when using `return` field on property
When the `return` field is used as the documentation body for
a property, we reset the source docstring to make sure that
the `return` field is also used as the summary. If we don't
do this, format_summary() will parse the source docstring
again, get an empty body again and produce an empty summary
as a result.

Ideally we'd summarize based on the parsed docstring instead
of the source string, but that requires significant changes
(see #86).