Support property setter and deleter methods, in a primitive way
Ideally we'd merge the setter and deleter documentation into the
property, but presenting that elegantly is going to take a lot
of work both to implement and evaluate.

For now, we rename the setters and deleters, so they show up in
the ouptut as separate functions, but don't overwrite the property
or cause warnings about duplicate functions.