Restore source link on package pages
The source link was missing since commit 5f83bb45 ("Remove
setSourceHref() call for package").

While there are separate Documentables for the package and the
`__init__.py`, the package docstring originates from `__init__.py`,
so that is where we should be linking to.

In the future we'll probably have a single Documentable, since
the language definition states that a package is in fact a special
kind of module.