Define version in metadata and extract it using importlib.metadata
Until now, pydoctor used Incremental to define the version in code
and extract it from there during the build to insert it into the
metadata. This commit reverses that approach, by defining the version
in metadata and having the code extract it from there.

Static metadata is easier to use by tools such as pydoctor itself.
In the future we'll probably want to adopt PEP 621 for that purpose;
this is a stepping stone.