Warn instead of crash when passing a non-class to classImplements()
The visit_Call_zope_interface_classImplements() method assumes that
if the first argument can be resolved to an object, that object will
be a class. Since that is the correct way to use classImplements(),
this assumption is usually true, but it doesn't have to hold when
analyzing broken code or mis-analyzing complex code.