Do not store decorators that we fail to analyze
If we don't have a name for the decorator, we can't do anything useful
with it, so there is no point in storing it. Not storing it means that
the code analyzing decorators later doesn't have to deal with 'base'
being None.

Note that analysis can fail in theory, but in practice I was unable
to find any decorators for which this happens: CPython's parser for
decorators seems to be quite strict and every attempt I made to
use a nonsensical decorator resulted in SyntaxError being raised.