Remove unused ZopeInterfaceClass.allImplementations property
This property was used by the pydoctor.html module, which was removed
over a year ago, in commit 8040e1d6. Since it did have test coverage,
it wasn't obvious that it was dead code.

After I added annotations, mypy reported that 'c' could be an instance
of ZopeInterfaceModule, which doesn't have a 'subclasses' attribute.
When I tried to find any use cases that could trigger this bug, I found
out that nothing uses this property.