Remove excessive caution from _EpydocReader.report()
After reading parts of the `docutils` code, I don't think it's likely
that `level` will be undefined or not an integer. For some messages
`line` could be undefined, but if it is defined it should be either
an integer or maybe `None`.

Nothing bad happens when generating API docs for CPython, which does
trigger lots of error reporting.

I also replaced the `> 2` magic value with `>= Reporter.ERROR_LEVEL`.
The value 2 is `WARNING_LEVEL`, but I thought it would be clearer to
check for `ERROR_LEVEL` (3) instead.