Support attr.ib definitions using positional arguments
Until now, we assumed that keyword arguments would be used. Examples
in the attrs documentation use keyword arguments exclusively, but the
API itself allows positional arguments as well.

I also added 'attr.attr' as a third full name for 'attr.ib': this
is deprecated in attrs, but might still be in use and it costs us
very little to support it.