Annotate methods of astbuilder's main node visitor
The added `expr is None` check in _handleDocstringUpdate() looks like
a bug fix, but it is not: in Python 3.8, ast.literal_eval() will raise
ValueError when it is passed a None expr, just like it does for actual
AST nodes that it cannot evaluate. This is not documented and seems
entirely accidental when reading the implementation, so it's better
to not rely on this. But it does mean that failing to handle None in
pydoctor couldn't lead to crashes.