Look up interface arguments in the right context
addInterfaceInfoToScope() would always look up names in the interface
arguments in the scope the interface was being added to, but this is
incorrect for both decorators and classImplements().

This inaccuracy was introduced for decorators in the previous commit,
when I moved the argument evaluation from visit_ClassDef() to
addInterfaceInfoToScope(). But for classImplements() the inaccuracy
was already there for a long time.