Improve lookup of implemented interfaces
addInterfaceInfoToScope() was split into two parts: one part that
records the interface's name in 'implements_directly' and one part
that records the 'implementedby_directly' reverse links and converts
non-interface classes to interfaces if necessary. The second part is
called after all modules have been processedm, to make sure we don't
miss interface relations when there are cyclic imports, which are
quite common in projects using type annotations.

The evaluation of interface arguments is done using node2fullname()
instead of astor, which makes it possible to tell apart non-name
expressions from names that cannot be found.

A new warning reports when an interface cannot be found and its name
does belong in the project we're documenting.

test_implementer_decoration_nonclass() didn't test anything that was
not tested by test_implementer_nonclass(), so I removed it.