Ignore docstring updates in functions
If an assignment to `__doc__` is inside a function, it probably does
at least one of two things:
- assign to an object determined by an argument
- assign a computed value

Either of these means that our simple static analysis has no chance
of determining how the docstring is updated. So instead of spewing
warnings, it's better to just ignore these.

This change avoids 8 warnings when processing Twisted.